Set some relationships laziness directly in models
There objects that are strictly related, like a vuln and its references,
or a workspace and its scope. They are queried together almost all the
times, so it is OK to define the relation in the model.
This is not the case for, for example, vulnerability and service since
it depends on the query.