Fix bug with pytest that was unexpectedly skipping some tests
Because of a bug with pytest, I can't simply mark TestListCommandView
with @pytest.mark.skip. I had to made it inherit from object instad of
ReadOnlyAPITests, and to manually skip the extra tests inside the class.
See https://docs.pytest.org/en/latest/skipping.html#skip-all-test-functions-of-a-class-or-module
and https://github.com/pytest-dev/pytest/issues/568 for more information