Set tool to imported commands after importing all the objects
The tool name is stored in CouchDB in every object's metadata instead of
in the CommandRunInformation object so I have to first create a command
with no tool, then import all the objects storing the tool and then
updating every command.

I also improved the logging of vuln template and user importers.