Improve status report API response time and command importer
I added command related properties to the model as column properties, so
I don't have to do this logic in the schema or in the view.
Now that I can undefer this field, it improves performances of the
status report API (in my workspace, by 4X).
Also improved the way of creating CommandObject's