Fix bug importing evidence from couchdb
When a file had a space, the URL to fetch it from couch was malformed
and it returned a 404 error code. This case wasn't handled properly so
the content of the imported evidence was a JSON 404 error message.

I fixed the URL to properly encode special characters and added an
assertion to check that the download request returned a 200 error code.