Improve migration check behavior
If the current revision does not exist in this branch, assume that this
is because a developer performed a migration in a new branch. In this
case, disable the migration check. This way, the dev doesn't need to
checkout to the new branch, perform a downgrade and checkout again to
the previous branch.