Fix CSRF vulnerability in all JSON endpoints
To fix the vuln, we need to use SameSite cookies. The upstream version
of Flask-KVSession doesn't support that, so we forked it, merged the
changes in https://github.com/mbr/flask-kvsession/pull/52 and uploaded
the new version to PyPI (with the name `flask-kvsession-fork`).

I also changed the session cookie name to discard old session cookies
that don't have the SameSite attribute unset.