Remove test_start_and_kill_faraday_server
It was not a unit test. It was created when we didn't have other
automated testing stuff other than the unit tests. Now that we have the
smoke test and the integration tests, test_start_and_kill_faraday_server
is redundant. Removing it will save 30 seconds in each tests run.